dom/inputmethod/mochitest/test_sendkey_cancel.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=952080
-->
<head>
  <title>SendKey with canceled keydown test for InputMethod API.</title>
  <script type="application/javascript;version=1.7" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.7" src="inputmethod_common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=952080">Mozilla Bug 952080</a>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.7">

// The input context.
var gContext = null;

inputmethod_setup(function() {
  runTest();
});

function runTest() {
  let im = navigator.mozInputMethod;

  im.oninputcontextchange = function() {
    ok(true, 'inputcontextchange event was fired.');
    im.oninputcontextchange = null;

    gContext = im.inputcontext;
    if (!gContext) {
      ok(false, 'Should have a non-null inputcontext.');
      inputmethod_cleanup();
      return;
    }

    test();
  };

  // Set current page as an input method.
  SpecialPowers.wrap(im).setActive(true);

  let iframe = document.createElement('iframe');
  iframe.src = 'file_test_sendkey_cancel.html';
  iframe.setAttribute('mozbrowser', true);
  document.body.appendChild(iframe);
}

function test() {
  gContext.sendKey(0, 'j', 0).then(function() {
    ok(false, 'sendKey was incorrectly resolved');

    inputmethod_cleanup();
  }, function(e) {
    ok(true, 'sendKey was rejected');

    inputmethod_cleanup();
  });
}

</script>
</pre>
</body>
</html>
